fix(snipe): Fix clean_edits breaking new events
`clean_edits` was resetting `self._deleted` and `self._message_history`
to a regular dict, which was breaking insertions by events if the
channel/message had not been added before.

Also fix that `self._message_history` had a deque as a default value
even though other places expected it to be a list due to mismatched type
annotations.

Signed-off-by: Leonhard Saam <leonhard.saam@yahoo.com>